requirements.txt
aiocache==0.12.2
aiofiles==0.8.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.3
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
easy-pil==0.3.3
fonttools==4.50.0
frozenlist==1.4.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
multidict==6.0.5
numpy==1.26.4
packaging==24.0
Pillow==10.1.0
ping3==4.0.5
psutil==5.9.8
py-cord==2.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-steam-api==1.2.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.8.0
update-checker==0.18.0
urllib3==2.2.1
yarl==1.9.4
